+++ app/javascript/artisans_ui/controllers/index.js
@@ -9,8 +9,10 @@ import AutogrowController from "./autogrow_controller"
 import BannerController from "./banner_controller"
 import CarouselController from "./carousel_controller"
 import ClipboardController from "./clipboard_controller"
+import ConfirmationController from "./confirmation_controller"
 import ModalController from "./modal_controller"
 import DropdownController from "./dropdown_controller"
+import SelectController from "./select_controller"
 import TooltipController from "./tooltip_controller"
 
 // Register controllers
@@ -20,8 +22,10 @@ application.register("artisans-ui--autogrow", AutogrowController)
 application.register("artisans-ui--banner", BannerController)
 application.register("artisans-ui--carousel", CarouselController)
 application.register("artisans-ui--clipboard", ClipboardController)
+application.register("artisans-ui--confirmation", ConfirmationController)
 application.register("artisans-ui--modal", ModalController)
 application.register("artisans-ui--dropdown", DropdownController)
+application.register("artisans-ui--select", SelectController)
 application.register("artisans-ui--tooltip", TooltipController)
 
-export { AccordionController, AnimatedNumberController, AutogrowController, BannerController, CarouselController, ClipboardController, ModalController, DropdownController, TooltipController }
+export { AccordionController, AnimatedNumberController, AutogrowController, BannerController, CarouselController, ClipboardController, ConfirmationController, ModalController, DropdownController, SelectController, TooltipController }